+++ src/main/java/madsciencemod/cards/GatherParts.java
@@ -18,7 +18,7 @@ public class GatherParts extends AbstractMadScienceCard {
     public static final String UPGRADE_DESCRIPTION = cardStrings.UPGRADE_DESCRIPTION;
     private static final int COST = 1;
     private static final int TRINKETS = 3;
-    private static final CardType TYPE = CardType.ATTACK;
+    private static final CardType TYPE = CardType.SKILL;
     private static final CardRarity RARITY = CardRarity.RARE;
     private static final CardTarget TARGET = CardTarget.NONE;