+++ src/main/java/madsciencemod/cards/SelfDestruct.java
@@ -20,7 +20,7 @@ public class SelfDestruct extends AbstractMadScienceCard {
     private static final int SELF_DMG = 6;
     private static final int ATTACK_DMG = 20;
     private static final int UPGRADE_ATTACK_DMG = 5;
-    private static final CardType TYPE = CardType.SKILL;
+    private static final CardType TYPE = CardType.ATTACK;
     private static final CardRarity RARITY = CardRarity.UNCOMMON;
     private static final CardTarget TARGET = CardTarget.ALL_ENEMY;