+++ src/main/java/madsciencemod/cards/HideBehindJunk.java
@@ -34,7 +34,7 @@ public class HideBehindJunk extends AbstractMadScienceCard {
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
         AbstractDungeon.actionManager.addToBottom(new GainBlockAction(p, p, this.block));
-        AbstractDungeon.actionManager.addToBottom(new ShuffleTrinketAction(this.magicNumber, true, true));
+        AbstractDungeon.actionManager.addToBottom(new ShuffleTrinketAction(this.magicNumber, false, true));
     }
 
     @Override