+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed: Self Destruct is now an Attack
 * Fixed: Infinite Journal relic description
 * UI: Added flavor text to relics
+* UI: Use custom mana symbol on cards. Requires BaseMod v2.16.0 / week 33 patch.
 
 ### v0.1.8
 * Changed Retarget: now gains 2 thorns per attack for 2(1) strengt, instead of 2(3) thorns for 1 strength.